Rename imessageForm to imessage in the component folder
