+++ index.message.js
@@ -2,7 +2,7 @@ import "./shim"; // eslint-disable-line
 import {
   AppRegistry,
 } from "react-native"; // eslint-disable-line
-import LiskMessageExtension from './src/components/imessageForm';
+import LiskMessageExtension from './src/components/imessage';
 
 AppRegistry.registerComponent(
   'LiskMessageExtension',
